requirements.txt
requests==2.32.5
python-vlc>=3.0.20123
mutagen>=1.47
pygame>=2.5
faster-whisper>=1.0